TimeoutWWW提交，需要实现事件通知时同时通知lua

+++ Assets/GameFramework/Scripts/Common/TimeOutWWW.cs
@@ -9,7 +9,10 @@ namespace GameFramework
 {
     public class TimeOutWWW : MonoBehaviour
     {
-        public delegate void WWWRstDel(double progress, int index, string msg);
+        public delegate void WWWRstDel(string noticeKey, double progress, int index, string msg);
+
+        public static string STR_SUCCEEDED = "succeeded";
+        public static string STR_FAILED = "failed";
 
         private WWW mWWW = null;
         private WWWRstDel mRstDel = null;
@@ -19,15 +22,18 @@ namespace GameFramework
         private bool mIsTimeOut;
         private long mTotalSize = 0;
         private long mDoneSize = 0;
-        private List<DownloadInfo> mList;
-        private List<DownloadInfo> mFialedList;
+        private List<WWWInfo> mList;
+        private List<WWWInfo> mFialedList;
         private int mDoneCount = 0;
+        private WWWType mType = WWWType.download;
+        private string mNoticeKey;
+
 
         #region MonoBehaviour
         void Awake()
         {
-            mList = new List<DownloadInfo>();
-            mFialedList = new List<DownloadInfo>();
+            mList = new List<WWWInfo>();
+            mFialedList = new List<WWWInfo>();
         }
 
         void Start()
@@ -55,24 +61,27 @@ namespace GameFramework
                 mWWW.Dispose();
                 mWWW = null;
             }
-
+            setLuaCallback(null);
         }
         #endregion MonoBehaviour
 
-        public void DownloadFile(DownloadInfo info, float timeoutSec, WWWRstDel del, LuaFunction lua)
+        public void DownloadFile(string noticeKey, WWWInfo info, float timeoutSec, WWWRstDel del, LuaFunction lua)
         {
+            mType = WWWType.download;
             mRstDel = del;
             setLuaCallback(lua);
             mTimeoutSec = timeoutSec;
             mTotalSize = info.Size;
+            mNoticeKey = noticeKey;
 
             mList.Clear();
             mList.Add(info);
         }
 
-        public void DownloadFiles(List<DownloadInfo> infos, float timeoutSec, WWWRstDel del, LuaFunction lua)
+        public void DownloadFiles(string noticeKey, List<WWWInfo> infos, float timeoutSec, WWWRstDel del, LuaFunction lua)
         {
-            if(null == infos || infos.Count == 0)
+            mType = WWWType.download;
+            if (null == infos || infos.Count == 0)
             {
                 callback(-1, "download files info list is null or empty!");
                 return;
@@ -80,6 +89,7 @@ namespace GameFramework
             mRstDel = del;
             setLuaCallback(lua);
             mTimeoutSec = timeoutSec;
+            mNoticeKey = noticeKey;
             for (int i = 0; i < infos.Count; i++)
             {
                 mTotalSize = infos[i].Size;
@@ -90,24 +100,50 @@ namespace GameFramework
             mFialedList.Clear();
         }
 
-        public void UploadFile()
+        public void UploadFile(string noticeKey, WWWInfo info, float timeoutSec, WWWRstDel del, LuaFunction lua)
         {
+            mType = WWWType.upload;
+            mRstDel = del;
+            setLuaCallback(lua);
+            mTimeoutSec = timeoutSec;
+            mTotalSize = info.Size;
+            mNoticeKey = noticeKey;
 
+            mList.Clear();
+            mList.Add(info);
         }
 
-        public void UploadFiles()
+        public void UploadFiles(string noticeKey, List<WWWInfo> infos, float timeoutSec, WWWRstDel del, LuaFunction lua)
         {
+            mType = WWWType.upload;
+            if (null == infos || infos.Count == 0)
+            {
+                callback(-1, "download files info list is null or empty!");
+                return;
+            }
+            mRstDel = del;
+            setLuaCallback(lua);
+            mTimeoutSec = timeoutSec;
+            mNoticeKey = noticeKey;
+            for (int i = 0; i < infos.Count; i++)
+            {
+                mTotalSize = infos[i].Size;
+            }
 
+            mList.Clear();
+            mList.AddRange(infos);
+            mFialedList.Clear();
         }
 
-        public void GetRequest()
+        public void RequestUrl(string noticeKey, string url, float timeoutSec, WWWRstDel del, LuaFunction lua)
         {
+            mType = WWWType.request;
+            mRstDel = del;
+            setLuaCallback(lua);
+            mTimeoutSec = timeoutSec;
+            mNoticeKey = noticeKey;
 
-        }
-
-        public void PostRequest()
-        {
-
+            mWWW = new WWW(url);
         }
         #region 私有方法
         private void setLuaCallback(LuaFunction lua)
@@ -124,11 +160,11 @@ namespace GameFramework
         {
             if (null != mRstDel)
             {
-                mRstDel(progress, mDoneCount, msg);
+                mRstDel(mNoticeKey, progress, mDoneCount, msg);
             }
             if (null != mLuaFunc)
             {
-                mLuaFunc.Call(progress, mDoneCount, msg);
+                mLuaFunc.Call(mNoticeKey, progress, mDoneCount, msg);
             }
             if (progress >= 1)
             {
@@ -138,26 +174,26 @@ namespace GameFramework
 
         private void clear()
         {
-            mRstDel = null;
-            mLuaFunc.Dispose();
-            mLuaFunc = null;
-            mTimeoutSec = 0f;
-            mTotalSize = 0;
-            if (null != mWWW)
-            {
-                mWWW.Dispose();
-                mWWW = null;
-            }
+            //mRstDel = null;
+            //mLuaFunc.Dispose();
+            //mLuaFunc = null;
+            //mTimeoutSec = 0f;
+            //mTotalSize = 0;
+            //if (null != mWWW)
+            //{
+            //    mWWW.Dispose();
+            //    mWWW = null;
+            //}
             if (null != mCoroutine)
             {
                 StopCoroutine(mCoroutine);
             }
+            Destroy(this, 2);
         }
 
         /// <summary>
         /// 检查www的超时
         /// </summary>
-        /// <param name="request"></param>
         /// <param name="timeout"></param>
         /// <returns></returns>
         private IEnumerator checkWWWTimeout(float timeout)
@@ -205,20 +241,44 @@ namespace GameFramework
         {
             for (int i = 0; i < mList.Count; i++)
             {
-                DownloadInfo info = mList[i];
+                WWWInfo info = mList[i];
                 if (string.IsNullOrEmpty(info.Url) || string.IsNullOrEmpty(info.TargetPath))
                 {
                     callback(-1f, "url/target is null or empty ");
                     continue;
                 }
-                mWWW = new WWW(info.Url);
+                if (WWWType.download == mType)
+                {
+                    mWWW = new WWW(info.Url);
+                }
+                if (WWWType.upload == mType)
+                {
+                    WWWUploadInfo uploadInfo = info as WWWUploadInfo;
+                    string path = uploadInfo.TargetPath;
+                    if (!string.IsNullOrEmpty(path) && Tools.CheckFileExists(path))
+                    {
+                        WWW file = new WWW(Tools.GetUrlPathWritebble(path));
+                        while (!file.isDone)
+                        {
+                            yield return null;
+                        }
+                        if (string.IsNullOrEmpty(file.error))
+                        {
+                            WWWForm form = new WWWForm();                                       //WWWForm是一个辅助类，该类用于生成表单数据，                                                                           //然后WWW类就可以将该表单数据post到web服务器上了
+                            form.AddBinaryData(uploadInfo.FieldName, file.bytes, uploadInfo.SaveName, uploadInfo.MimeType);         //添加二进制文件到表单，使用该函数可以上传文件或者图片到Web服务器     
+                            mWWW = new WWW(uploadInfo.Url, form);
+
+                        }
+                    }
+                }
+
                 if (null != mWWW)
                 {
                     callback(-1f, "www is not null");
                     continue;
                 }
                 yield return checkWWWTimeout(mTimeoutSec);
-                if(mIsTimeOut)
+                if (mIsTimeOut)
                 {
                     mWWW.Dispose();
                     mWWW = null;
@@ -231,28 +291,34 @@ namespace GameFramework
                     {
                         if (string.IsNullOrEmpty(mWWW.error))
                         {
-                            //下载失败
+                            //WWW上传或下载失败
+                            mWWW.Dispose();
+                            mWWW = null;
                             callback(-1f, mWWW.error);
                             mFialedList.Add(info);
                         }
                         else
                         {
-                            //下载完成
                             mDoneSize += info.Size;
                             mDoneCount += 1;
-                            string savePath = info.TargetPath;
-                            if (!string.IsNullOrEmpty(savePath))
+                            if (WWWType.download == mType)
                             {
-                                if (File.Exists(savePath))
+                                //下载完成,拷贝到目标位置
+                                string savePath = info.TargetPath;
+                                if (!string.IsNullOrEmpty(savePath))
                                 {
-                                    File.Delete(savePath);
+                                    if (File.Exists(savePath))
+                                    {
+                                        File.Delete(savePath);
+                                    }
+                                    Tools.CheckDirExists(Directory.GetParent(savePath).ToString(), true);
+                                    FileStream fsDes = File.Create(savePath);
+                                    fsDes.Write(mWWW.bytes, 0, mWWW.bytes.Length);
+                                    fsDes.Flush();
+                                    fsDes.Close();
                                 }
-                                Tools.CheckDirExists(Directory.GetParent(savePath).ToString(), true);
-                                FileStream fsDes = File.Create(savePath);
-                                fsDes.Write(mWWW.bytes, 0, mWWW.bytes.Length);
-                                fsDes.Flush();
-                                fsDes.Close();
                             }
+                                
                             callback(getProgress(0), "done");
                         }
                     }
@@ -263,13 +329,12 @@ namespace GameFramework
             yield return null;
             if (mFialedList.Count == 0)
             {
-                callback(1, "succeeded");
+                callback(1, STR_SUCCEEDED);
             }
             else
             {
-                callback(-1, "failed");
+                callback(1, STR_FAILED);
             }
-            Destroy(this);
         }
 
         private double getProgress(int curSize)
@@ -278,7 +343,7 @@ namespace GameFramework
             if (mTotalSize > 0)
             {
                 progress = (mDoneSize + curSize) / (double)mTotalSize;
-                
+
                 return progress;
             }
             else
@@ -301,26 +366,230 @@ namespace GameFramework
                 {
                     if (string.IsNullOrEmpty(mWWW.error))
                     {
-                        //下载失败
-                        //TODO:
+                        //请求失败
+                        callback(1, STR_FAILED);
                     }
                     else
                     {
-                        //下载完成
-                        //TODO:
+                        //请求完成
+                        callback(1, STR_SUCCEEDED);
                     }
                 }
-                yield return null;
-                Destroy(this);
             }
         }
         #endregion 私有方法
     }
 
-    public struct DownloadInfo
+    public class WWWInfo : IDisposable
     {
         public string Url;
         public string TargetPath;
         public long Size;
+        public Dictionary<string, string> Headers;
+
+        public WWWInfo()
+        {
+
+        }
+
+        public WWWInfo(string url, string targetPath, long size=0, Dictionary<string, string> headers = null)
+        {
+            Url = url;
+            TargetPath = targetPath;
+            Size = size;
+            Headers = headers;
+        }
+
+        public WWWInfo(LuaTable table)
+        {
+            if (null != table)
+            {
+                Url = table.RawGet<string, string>("url");
+                TargetPath = table.RawGet<string, string>("targetPath");
+                Size = table.RawGet<string, long>("size");
+                LuaTable ht = table.RawGet<string, LuaTable>("headers");
+                if(null != ht)
+                {
+                    Headers = ht.ToDictTable<string, string>().ToDictionary();
+                    ht.Dispose();
+                    ht = null;
+                }
+                table.Dispose();
+                table = null;
+            }
+        }
+
+        public static List<WWWInfo> GetListByLua(LuaTable table)
+        {
+            List<WWWInfo> infos = new List<WWWInfo>();
+            LuaArrayTable luaArray = new LuaArrayTable(table);
+            luaArray.ForEach((obj) =>
+            {
+                LuaTable t = obj as LuaTable;
+                if (null != t)
+                {
+                    infos.Add(new WWWInfo(t));
+                    t.Dispose();
+                    t = null;
+                }
+            });
+            luaArray.Dispose();
+            luaArray = null;
+            table.Dispose();
+            table = null;
+            return infos;
+        }
+
+        public void Dispose()
+        {
+            Headers.Clear();
+        }
+    }
+
+    public class WWWUploadInfo : WWWInfo
+    {
+        public string SaveName;
+        public string FieldName;
+        public string MimeType;
+
+        public WWWUploadInfo()
+        {
+        }
+
+        public WWWUploadInfo(string url, string targetPath, string fieldName, string saveName="", string mimeType = "", long size = 0, Dictionary<string, string> headers = null) : base(url, targetPath, size, headers)
+        {
+            FieldName = fieldName;
+            SaveName = saveName;
+            if(string.IsNullOrEmpty(SaveName))
+            {
+                SaveName = Tools.FormatPathStr(targetPath);
+                SaveName = SaveName.Substring(SaveName.LastIndexOf("/", StringComparison.Ordinal)+1);
+            }
+            MimeType = mimeType;
+            if (string.IsNullOrEmpty(MimeType))
+            {
+                MimeType = GetMimeTypeByName(saveName);
+            }
+        }
+
+        public WWWUploadInfo(LuaTable table)
+        {
+            if (null != table)
+            {
+                Url = table.RawGet<string, string>("url");
+                TargetPath = table.RawGet<string, string>("targetPath");
+                Size = table.RawGet<string, long>("size");
+                SaveName = table.RawGet<string, string>("saveName");
+                FieldName = table.RawGet<string, string>("fieldName");
+                MimeType = table.RawGet<string, string>("mimeType");
+                LuaTable ht = table.RawGet<string, LuaTable>("headers");
+                if (null != ht)
+                {
+                    Headers = ht.ToDictTable<string, string>().ToDictionary();
+                    ht.Dispose();
+                    ht = null;
+                }
+                table.Dispose();
+                table = null;
+            }
+        }
+
+        public static new List<WWWInfo> GetListByLua(LuaTable table)
+        {
+            List<WWWInfo> infos = new List<WWWInfo>();
+            LuaArrayTable luaArray = new LuaArrayTable(table);
+            luaArray.ForEach((obj) =>
+            {
+                LuaTable t = obj as LuaTable;
+                if (null != t)
+                {
+                    infos.Add(new WWWUploadInfo(t));
+                    t.Dispose();
+                    t = null;
+                }
+            });
+            luaArray.Dispose();
+            luaArray = null;
+            table.Dispose();
+            table = null;
+            return infos;
+        }
+
+        public static string GetMimeTypeByName(string saveName)
+        {
+            #region GetMimeTypeByName
+            string type = "application/octet-stream";
+            string extension = Path.GetExtension(saveName);
+            switch(extension)
+            {
+                case ".swf":
+                    return "application/x-shockwave-flash";
+                case ".dll":
+                    return "application/x-msdownload";
+                case ".exe":
+                    return "application/octet-stream";
+                case ".rar":
+                    return "application/octet-stream";
+                case ".tar":
+                    return "application/x-tar";
+                case ".tgz":
+                    return "application/x-compressed";
+                case ".zip":
+                    return "application/x-zip-compressed";
+                case ".z":
+                    return "application/x-compress";
+                case ".wav":
+                    return "audio/wav";
+                case ".wma":
+                    return "audio/x-ms-wma";
+                case ".wmv":
+                    return "video/x-ms-wmv";
+                case ".mp3":
+                case ".mp2":
+                case ".mpe":
+                case ".mpeg":
+                case ".mpg":
+                    return "audio/mpeg";
+                case ".rm":
+                    return "application/vnd.rn-realmedia";
+                case ".mid":
+                case ".midi":
+                case ".rmi":
+                    return "audio/mid";
+                case ".bmp":
+                    return "image/bmp";
+                case ".gif":
+                    return "image/gif";
+                case ".png":
+                    return "image/png";
+                case ".tif":
+                case ".tiff":
+                    return "image/tiff";
+                case ".jpe":
+                case ".jpeg":
+                case ".jpg":
+                    return "image/jpeg";
+                case ".txt":
+                case ".log":
+                    return "text/plain";
+                case ".xml":
+                    return "text/xml";
+                case ".html":
+                    return "text/html";
+                case ".css":
+                    return "text/css";
+                case ".js":
+                    return "text/javascript";
+            }
+            return type;
+            #endregion GetMimeTypeByName
+        }
+    }
+
+    public enum WWWType
+    {
+        download,
+        upload,
+        request,
     }
 }
